Removing osrm from turn generation.

+++ routing/road_graph.cpp
@@ -167,7 +167,7 @@ void IRoadGraph::CrossOutgoingLoader::LoadEdges(FeatureID const & featureId, Roa
   ForEachEdge(roadInfo, [&featureId, &roadInfo, this](size_t segId, Junction const & endJunction,
                                                       bool forward) {
     if (forward || roadInfo.m_bidirectional || m_mode == IRoadGraph::Mode::IgnoreOnewayTag)
-      m_edges.push_back(Edge::MakeReal(featureId, forward, segId, m_cross, endJunction));
+      m_edges.push_back(Edge::MakeReal(featureId, forward, static_cast<uint32_t>(segId), m_cross, endJunction));
   });
 }
 
@@ -177,7 +177,7 @@ void IRoadGraph::CrossIngoingLoader::LoadEdges(FeatureID const & featureId, Road
   ForEachEdge(roadInfo, [&featureId, &roadInfo, this](size_t segId, Junction const & endJunction,
                                                       bool forward) {
     if (!forward || roadInfo.m_bidirectional || m_mode == IRoadGraph::Mode::IgnoreOnewayTag)
-      m_edges.push_back(Edge::MakeReal(featureId, !forward, segId, endJunction, m_cross));
+      m_edges.push_back(Edge::MakeReal(featureId, !forward, static_cast<uint32_t>(segId), endJunction, m_cross));
   });
 }
